--- alembic/data_schema/versions/0f3fded53069_.py
@@ -1,95 +1,0 @@
-"""empty message
-
-Revision ID: 0f3fded53069
-Revises: 54adea236048
-Create Date: 2025-02-20 15:13:33.457997
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = "0f3fded53069"
-down_revision: Union[str, None] = "54adea236048"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column("stock_indices_1m", "date", existing_type=mysql.DATETIME(), nullable=True, existing_comment="날짜")
-    op.alter_column(
-        "stock_information",
-        "is_activate",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=None,
-        existing_comment="사용 가능 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_trend",
-        "is_trading_stopped",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=None,
-        existing_comment="거래 정지 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_trend",
-        "is_delisted",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=None,
-        existing_comment="상장 폐지 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_trend",
-        "is_activate",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=None,
-        existing_comment="사용 가능 여부",
-        existing_nullable=True,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "stock_trend",
-        "is_activate",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=sa.text("'1'"),
-        existing_comment="사용 가능 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_trend",
-        "is_delisted",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=sa.text("'0'"),
-        existing_comment="상장 폐지 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_trend",
-        "is_trading_stopped",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=sa.text("'0'"),
-        existing_comment="거래 정지 여부",
-        existing_nullable=True,
-    )
-    op.alter_column(
-        "stock_information",
-        "is_activate",
-        existing_type=mysql.TINYINT(display_width=1),
-        server_default=sa.text("'1'"),
-        existing_comment="사용 가능 여부",
-        existing_nullable=True,
-    )
-    op.alter_column("stock_indices_1m", "date", existing_type=mysql.DATETIME(), nullable=False, existing_comment="날짜")
-    # ### end Alembic commands ###
