+++ alembic/data_schema/versions/2c3dda97574f_v55.py
@@ -1,0 +1,32 @@
+"""v55
+
+Revision ID: 2c3dda97574f
+Revises: 54adea236048
+Create Date: 2025-02-21 15:32:46.574397
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+# revision identifiers, used by Alembic.
+revision: str = "2c3dda97574f"
+down_revision: Union[str, None] = "54adea236048"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "stock_information", sa.Column("is_pub", sa.Boolean(), nullable=True, comment="is_publishable, 수집 가능 여부")
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("stock_information", "is_pub")
+    # ### end Alembic commands ###
